GG+VDP fixes
- fixed GG i/o register on port 5 (returns 0 instead of ff)
- fixed vcounter values
- improved hcounter implementation, although it still needs work
- tweaked audio buffering to improve smoothness of sound

+++ vdp/masterSystem/vdpMS.h
@@ -129,7 +129,7 @@ private:
     uint8_t line_int_cur; // Current value of line interrupt counter
     bool line_int_active;
     bool scr_int_active;
-    uint8_t curLine; //vcounter
+    unsigned int curLine; //vcounter
     uint8_t latchedPixel; //hcounter
     uint8_t fifthSprite; //fifth sprite found in SG-1000 mode
     bool sprOverflow;    //5 sprites found in SG-1000, 8 sprites found in SMS
